requirements.txt
fastapi
uvicorn
PyMuPDF
python-multipart
openai
python-dotenv
